+++ src/V1/Tests/TestFiles/ObjectTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using ServiceBricks.Cache;
 using ServiceBricks.Cache.EntityFrameworkCore;
 
@@ -17,7 +18,7 @@ namespace ServiceBricks.Xunit
         }
 
         [Fact]
-        public virtual async Task ExpirationTaskTests()
+        public virtual async Task ExpirationTaskQueueWorkTests()
         {
             var loggerFactory = SystemManager.ServiceProvider.GetRequiredService<ILoggerFactory>();
             var taskQueue = SystemManager.ServiceProvider.GetRequiredService<ITaskQueue>();
@@ -30,27 +31,65 @@ namespace ServiceBricks.Xunit
                 ExpirationDate = DateTimeOffset.UtcNow.AddSeconds(-1)
             };
 
-            var respcreate = apiservice.Create(startdto);
+            var respcreate = await apiservice.CreateAsync(startdto);
             Assert.True(respcreate.Success);
             Assert.True(respcreate.Item != null);
 
-            var respg = apiservice.Get(respcreate.Item.StorageKey);
+            var respg = await apiservice.GetAsync(respcreate.Item.StorageKey);
             Assert.True(respg.Success);
             Assert.True(respg.Item != null);
 
             CacheExpirationTask.QueueWork(taskQueue);
+
+            CancellationTokenSource cts = new CancellationTokenSource();
+            cts.CancelAfter(5000);
+            while (!cts.Token.IsCancellationRequested)
+            {
+                var respTest = await apiservice.GetAsync(respcreate.Item.StorageKey);
+                if (respTest.Item == null)
+                    break;
+            }
+
+            respg = await apiservice.GetAsync(respcreate.Item.StorageKey);
+
+            Assert.True(respg.Item == null);
+        }
+
+        [Fact]
+        public virtual async Task ExpirationTaskMethodTests()
+        {
+            var loggerFactory = SystemManager.ServiceProvider.GetRequiredService<ILoggerFactory>();
+            var taskQueue = SystemManager.ServiceProvider.GetRequiredService<ITaskQueue>();
+
+            var apiservice = SystemManager.ServiceProvider.GetRequiredService<ICacheDataApiService>();
+
+            var startdto = new CacheDataDto()
+            {
+                CacheKey = Guid.NewGuid().ToString(),
+                ExpirationDate = DateTimeOffset.UtcNow.AddSeconds(-1)
+            };
+
+            var respcreate = await apiservice.CreateAsync(startdto);
+            Assert.True(respcreate.Success);
+            Assert.True(respcreate.Item != null);
+
+            var respg = await apiservice.GetAsync(respcreate.Item.StorageKey);
+            Assert.True(respg.Success);
+            Assert.True(respg.Item != null);
+
             CacheExpirationTask.Worker worker = new CacheExpirationTask.Worker(apiservice);
             await worker.DoWork(new CacheExpirationTask.Detail(), CancellationToken.None);
 
             CancellationTokenSource cts = new CancellationTokenSource();
-            cts.CancelAfter(3000);
-            while (apiservice.Get(respcreate.Item.StorageKey).Item != null)
+            cts.CancelAfter(5000);
+            while (!cts.Token.IsCancellationRequested)
             {
-                if (cts.Token.IsCancellationRequested)
+                var respTest = await apiservice.GetAsync(respcreate.Item.StorageKey);
+                if (respTest.Item == null)
                     break;
             }
 
-            respg = apiservice.Get(respcreate.Item.StorageKey);
+            respg = await apiservice.GetAsync(respcreate.Item.StorageKey);
 
             Assert.True(respg.Item == null);
         }
@@ -60,12 +99,13 @@ namespace ServiceBricks.Xunit
         {
             var loggerFactory = SystemManager.ServiceProvider.GetRequiredService<ILoggerFactory>();
             var taskQueue = SystemManager.ServiceProvider.GetRequiredService<ITaskQueue>();
-
             var apiservice = SystemManager.ServiceProvider.GetRequiredService<ICacheDataApiService>();
+            var semaphoreOptions = new OptionsWrapper<SemaphoreOptions>(new SemaphoreOptions());
 
             var timer = new CacheExpirationTimer(
                 SystemManager.ServiceProvider,
-                SystemManager.ServiceProvider.GetRequiredService<ILoggerFactory>());
+                SystemManager.ServiceProvider.GetRequiredService<ILoggerFactory>(),
+                semaphoreOptions);
 
             var startdto = new CacheDataDto()
             {
@@ -73,24 +113,24 @@ namespace ServiceBricks.Xunit
                 ExpirationDate = DateTimeOffset.UtcNow.AddSeconds(-1)
             };
 
-            var respcreate = apiservice.Create(startdto);
+            var respcreate = await apiservice.CreateAsync(startdto);
             Assert.True(respcreate.Success);
             Assert.True(respcreate.Item != null);
 
             await timer.StartAsync(CancellationToken.None);
 
             CancellationTokenSource cts = new CancellationTokenSource();
-            cts.CancelAfter(3000);
-            while (apiservice.Get(respcreate.Item.StorageKey).Item != null)
+            cts.CancelAfter(5000);
+            while (!cts.Token.IsCancellationRequested)
             {
-                if (cts.Token.IsCancellationRequested)
+                var respTest = await apiservice.GetAsync(respcreate.Item.StorageKey);
+                if (respTest.Item == null)
                     break;
-                await Task.Delay(100);
             }
 
             await timer.StopAsync(CancellationToken.None);
 
-            var respg = apiservice.Get(respcreate.Item.StorageKey);
+            var respg = await apiservice.GetAsync(respcreate.Item.StorageKey);
             Assert.True(respg.Item == null);
         }
 
@@ -111,22 +151,22 @@ namespace ServiceBricks.Xunit
             Assert.True(result);
 
             // Get
-            var respGet = apiservice.Get(keyName);
+            var respGet = await apiservice.GetAsync(keyName);
             Assert.True(respGet.Success);
             Assert.True(respGet.Item != null);
 
             // Update
             respGet.Item.CacheValue = Guid.NewGuid().ToString();
-            var respUpdate = apiservice.Update(respGet.Item);
+            var respUpdate = await apiservice.UpdateAsync(respGet.Item);
 
             result = await service.ShouldThisServerRunForProcessAsync("test");
             Assert.True(!result); // should not
 
-            // take over process
+            // take over process (storage repository skips all business rules)
             var storagerepo = SystemManager.ServiceProvider.GetRequiredService<IStorageRepository<CacheData>>();
             var respitem = storagerepo.Get(new CacheData() { CacheKey = keyName });
             respitem.Item.UpdateDate = DateTimeOffset.UtcNow.Subtract(service.HeartbeatTimeout);
-            storagerepo.Update(respitem.Item);
+            await storagerepo.UpdateAsync(respitem.Item);
 
             result = await service.ShouldThisServerRunForProcessAsync("test");
             Assert.True(result); // take over
@@ -159,6 +199,16 @@ namespace ServiceBricks.Xunit
             IConfiguration config = new ConfigurationBuilder().Build();
 
             services.AddServiceBricksCacheClient(config);
+            return Task.CompletedTask;
+        }
+
+        [Fact]
+        public virtual Task AddCacheClientForServiceTests()
+        {
+            IServiceCollection services = new ServiceCollection();
+            IConfiguration config = new ConfigurationBuilder().Build();
+
+            services.AddServiceBricksCacheClientForService(config);
             return Task.CompletedTask;
         }
     }